+++ CHANGES.rst
@@ -1,3 +1,9 @@
+Release 0.0.8
+=============
+
+* **MinHashEncoder**: Added a "handle_missing" attribute to allow encoding
+  with missing values.
+
 Release 0.0.7
 =============